+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 4.2.5
+  - Added optional parameter `disableBlink` to [ThemedTimePicker] and [ThemedTimeRangePicker] to disable blinking of the inputs in the dialog
+  - Fixed a bug on the [ThemedTimePicker] and [ThemedTimeRangePicker] dialog in which the increase and decrease buttons were not updating the state of the widget
+
+## 4.2.4
+- Removed paginator from [ThemedTable<T>] in mobile mode.
+
 ## 4.2.3
 - Major change in [ThemedTable<T>] to handle dynamic columns and rows. Also, changes to horizontal scrolling inside of the table.